implemented coments
